mstrando resultado da soma no html

+++ produt.js
@@ -93,9 +93,9 @@ class Produto {
 
         
             for (let i = 0; i < this.arrayProdutos.length; i++) {
-            total = parseFloat(total) + parseFloat(this.arrayProdutos[i]['nomeValor']);
+            total = parseFloat(total) + parseInt(this.arrayProdutos[i]['nomeValor']);
 
-      document.getElementById("result").innerHTML = "total:"+total;
+      document.getElementById("result").innerHTML = "total:" +total;
